Started combining chapters, fixing references.
I added a custom stylesheet, thesistools.sty, and a bibliography merge
script, mergeBibs.sh, based on BibTools for merging all the
bibliographies into one. I added multiple stylesheets from CTAN that are
not found in the RHEL7 yum repositories. A lot of the other updates
include things that I had accumulated but not committed over time.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>